ssize_t redefinition on Windows - int/int64 vs long/long long (#1337)
* ssize_t redefinition on Windows - int/int64 vs long/long long

* Define ssize_t as __int64 for _WIN64, not long long

+++ httplib.h
@@ -132,7 +132,7 @@
 #ifdef _WIN64
 using ssize_t = __int64;
 #else
-using ssize_t = int;
+using ssize_t = long;
 #endif
 #endif // _MSC_VER
